requirements.txt
streamlit==1.24.1
pandas
openpyxl==3.1.2
sqlalchemy==1.4.46
bcrypt==4.0.1
streamlit-authenticator
PyYAML
email-validator==1.3.1
python-dotenv==0.21.0